+++ inc/hdp_math_utils.h
@@ -9,24 +9,24 @@ SumOfLogsMemo* new_log_sum_memo();
 void destroy_log_sum_memo(SumOfLogsMemo* memo);
 
 // returns log(Gamma(n / 2)) in amortized constant time with low risk of overflow
-double log_gamma_half(int n, SumOfLogsMemo* sum_of_logs_memo);
-double sum_of_logs(SumOfLogsMemo* memo, int n);
+double log_gamma_half(int64_t n, SumOfLogsMemo* sum_of_logs_memo);
+double sum_of_logs(SumOfLogsMemo* memo, int64_t n);
 
 // returns log(x + y) without leaving log transformed space
 double add_logs(double log_x, double log_y);
 // quick-select on array copy (does not alter original array)
-double median(double* arr, int length);
-double max(double* arr, int length);
+double median(double* arr, int64_t length);
+double max(double* arr, int64_t length);
 
 // returns the index of the first element of arr greater or equal to x, assuming arr is sorted
 // returns final index if x is greater than all elements of arr
-int bisect_left(double x, double* arr, int length);
+int64_t bisect_left(double x, double* arr, int64_t length);
 
-double* spline_knot_slopes(double* x, double* y, int length);
-double spline_interp(double query_x, double* x, double* y, double* slope, int length);
-double grid_spline_interp(double query_x, double* x, double* y, double* slope, int length);
+double* spline_knot_slopes(double* x, double* y, int64_t length);
+double spline_interp(double query_x, double* x, double* y, double* slope, int64_t length);
+double grid_spline_interp(double query_x, double* x, double* y, double* slope, int64_t length);
 
-double* linspace(double start, double stop, int length);
+double* linspace(double start, double stop, int64_t length);
 
 double rand_uniform(double a);
 double rand_beta(double a, double b);
